Fix Linux build

+++ CollisionGrid/PlatformTypes.h
@@ -18,7 +18,7 @@
 // Unsigned base types.
 typedef unsigned char 		uint8;		// 8-bit  unsigned.
 typedef unsigned short int	uint16;		// 16-bit unsigned.
-typedef unsigned long		uint32;		// 32-bit unsigned.
+typedef DWORD			uint32;		// 32-bit unsigned.
 typedef unsigned long long	uint64;		// 64-bit unsigned.
 
 // Signed base types.
